+++ changes.md
@@ -1,3 +1,7 @@
+#### 1.0.3
+* Improved the way hostnames are handled and added to containers. 
+* Setup Docker Image name and ID to be stored in git config after a container starts in [docker.image.name] and [docker.image.id].
+
 #### 1.0.2
 * Added [git docker clean] command to remove all Git Config settings.
 * Added automatic removal of .git from remote URL.